+++ bycon/beaconServer/configuration.py
@@ -20,14 +20,9 @@ def configuration():
     initialize_bycon_service(byc, "configuration")
     r = BeaconInfoResponse(byc)
     c_f = get_schema_file_path(byc, "beaconConfiguration")
-    c = load_yaml_empty_fallback( c_f )
+    c = load_yaml_empty_fallback(c_f)
+    print_json_response(r.populatedInfoResponse(c), byc["env"])
 
-    byc.update({
-        "service_response": r.populatedInfoResponse(c),
-        "error_response": r.errorResponse()
-    })
-
-    cgi_print_response( byc, 200 )
 
 ################################################################################
 ################################################################################