+++ bycon/lib/query_generation.py
@@ -44,6 +44,11 @@ class ByconQuery():
         self.debug_mode = byc.get("debug_mode", False)
         self.test_mode = byc.get("test_mode", False)
         self.test_mode_count = int(byc.get('test_mode_count', 5))
+        self.defaults = byc.get("beacon_defaults", {})
+        self.response_types = self.defaults.get("entity_defaults", {})
+
+        f_t_d = self.response_types.get("filteringTerm", {})
+        self.filtering_terms_coll = f_t_d.get("collection", "___none___")
 
         if dataset_id is False:
             self.ds_id = byc.get("dataset_ids", False)[0]
@@ -53,40 +58,27 @@ class ByconQuery():
         self.arguments = byc.get("form_data", {})
         self.argument_definitions = byc.get("argument_definitions", {})
         self.filters = byc.get("filters", [])
-        self.filtering_terms_coll = byc.get("filtering_terms_coll", "___none___")
-
-        self.dbpars = {
-            "mongohost": byc.get("mongohost", "localhost"),
-            "services_db": byc.get("services_db", "___none___"),
-            "genes_coll": byc.get("genes_coll", "___none___")
-        }
 
         self.requested_entity = byc.get("request_entity_id", False)
         self.response_entity = byc.get("response_entity_id", False)
         self.path_id_value = byc.get("request_entity_path_id_value", False)
-
-        self.defaults = byc.get("beacon_defaults", {})
 
         self.variant_request_type = byc.get("variant_request_type", "___none___")
         self.variant_request_definitions = byc.get("variant_request_definitions", {})
         self.varguments = byc.get("varguments", {})
 
         self.filter_definitions = byc.get("filter_definitions", {})
-        ff = byc.get("filter_flags", {})
 
+        ff = byc.get("filter_flags", {})
         self.filter_descendants = ff.get("descendants", True)
         self.filter_logic = ff.get("logic", '$and')
 
-        self.housekeeping_db = byc.get("housekeeping_db", "___none___")
-        self.handover_coll = byc.get("handover_coll", "___none___")
-        self.services_db = byc.get("services_db", "___none___")
-        self.genes_coll = byc.get("genes_coll", "___none___")
+        self.db_config = byc.get("db_config", {})
 
         pagination = byc.get("pagination", {"skip": 0, "limit": 0})
         self.limit = pagination.get("limit", 0)
         self.skip = pagination.get("skip", 0)
 
-        self.response_types = self.defaults.get("entity_defaults")
 
         self.queries = {
             "expand": True,
@@ -210,7 +202,11 @@ class ByconQuery():
         if not r_c:
             return
 
-        data_db = MongoClient(host=self.dbpars["mongohost"])[self.ds_id]
+        mdb_c = self.db_config
+        db_host = mdb_c.get("host", "localhost")
+
+
+        data_db = MongoClient(host=db_host)[self.ds_id]
         data_collnames = data_db.list_collection_names()
 
         if r_c not in data_collnames:
@@ -279,7 +275,7 @@ class ByconQuery():
         # query database for gene and use coordinates to create range query
         vp = self.varguments
 
-        gene_data, e = GeneInfo(self.dbpars).returnGene(vp["gene_id"])
+        gene_data, e = GeneInfo(self.db_config).returnGene(vp["gene_id"])
 
         # TODO: error report/warning
         if not gene_data:
@@ -441,7 +437,10 @@ class ByconQuery():
         if len(self.filters) < 1:
             return
 
-        data_db = MongoClient(host=self.dbpars["mongohost"])[self.ds_id]
+        mdb_c = self.db_config
+        db_host = mdb_c.get("host", "localhost")
+
+        data_db = MongoClient(host=db_host)[self.ds_id]
         coll_coll = data_db[ self.filtering_terms_coll ]
         self.collation_ids = coll_coll.distinct("id", {})
 
@@ -463,6 +462,7 @@ class ByconQuery():
             f_info = self.__query_from_collationed_filter(coll_coll, f_val)
             if f_info is False:
                 f_info = self.__query_from_filter_definitions(f_val)
+
             if f_info is False:
                 continue
 
@@ -501,7 +501,6 @@ class ByconQuery():
                     f_lists[f_entity][f_field].append(f_info["id"])
 
         # now processing the filter lists into the queries
-
         for f_entity in f_lists.keys():
             f_s_l = []
             for f_field, f_query_vals in f_lists[f_entity].items():
@@ -572,7 +571,7 @@ class ByconQuery():
                 #     response_add_filter_warnings(self.byc, ftw)
                 return f_info
 
-        return f_info
+        return False
 
 
     # -------------------------------------------------------------------------#
@@ -600,9 +599,13 @@ class ByconQuery():
         if not accessid:
             return
 
-        ho_client = MongoClient(host=self.dbpars["mongohost"])
-        ho_db = ho_client[self.housekeeping_db]
-        ho_coll = ho_db[self.handover_coll]
+        mdb_c = self.db_config
+        db_host = mdb_c.get("host", "localhost")
+        ho_dbname = mdb_c.get("housekeeping_db", "___none___")
+        ho_collname = mdb_c.get("handover_coll", "___none___")
+
+        ho_client = MongoClient(host=db_host)
+        ho_coll = ho_client[ho_dbname].ho_db[ho_collname]
         h_o = ho_coll.find_one({"id": accessid})
 
         # accessid overrides ... ?
@@ -626,7 +629,6 @@ class ByconQuery():
     # -------------------------------------------------------------------------#
 
     def __update_queries_for_entity(self, query, entity):
-
         logic = self.filter_logic
         r_t_s = self.response_types
         r_c = r_t_s[entity].get("collection")